+++ src/server/.babelrc.js
@@ -4,7 +4,7 @@ module.exports = {
     [
       '@babel/env',
       {
-        targets: 'node 8.5',
+        targets: 'node >= 8.5',
         modules: process.env.BABEL_ESM ? false : 'commonjs',
         shippedProposals: true,
         loose: true