+++ src/universal/.babelrc.js
@@ -5,7 +5,7 @@ module.exports = {
     [
       '@babel/env',
       {
-        targets: '> 0.5%, not dead, node 8.5',
+        targets: '> 0.5%, not dead, node >= 8.5',
         modules: process.env.BABEL_ESM ? false : 'commonjs',
         shippedProposals: true,
         loose: true