+++ changelog.md
@@ -8,6 +8,7 @@
 - Test `GraphQL.operate()` with both `reloadOnLoad` and `resetOnLoad` options true.
 - Use string `FormData` field names, as some `FormData` polyfills don't coerce numbers like native implementations do.
 - Test files in variables result in appropriate fetch options for a valid [GraphQL multipart request](https://github.com/jaydenseric/graphql-multipart-request-spec).
+- Tidied test names.
 
 ## 8.2.0